requirements/frozen.py3.txt
# this requirement file is autogenerated, please run requirements/freeze_all.py to upgrade correctly
# currently requirements/freeze_all.py needs to be executed for both python2 and python3 !
# to upgrade a single package use requirements/freeze_all.py --upgrade-only apackage


adal==1.2.1
aenum==2.1.2
alabaster==0.7.12
ansible==2.7.10
anytree==2.6.0
apipkg==1.5
appdirs==1.4.3
asn1crypto==0.24.0
attrs==19.1.0
azure==4.0.0
azure-applicationinsights==0.1.0
azure-batch==4.1.3
azure-common==1.1.18
azure-cosmosdb-table==1.0.5
azure-datalake-store==0.0.44
azure-eventgrid==1.2.0
azure-graphrbac==0.40.0
azure-keyvault==1.1.0
azure-loganalytics==0.1.0
azure-mgmt==4.0.0
azure-mgmt-advisor==1.0.1
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.50.0
azure-mgmt-batch==5.0.1
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-cdn==3.1.0
azure-mgmt-cognitiveservices==3.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-compute==4.6.0
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.4.1
azure-mgmt-containerregistry==2.7.0
azure-mgmt-containerservice==4.4.0
azure-mgmt-cosmosdb==0.4.1
azure-mgmt-datafactory==0.6.0
azure-mgmt-datalake-analytics==0.6.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==1.0.0
azure-mgmt-devspaces==0.1.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==1.0.0
azure-mgmt-eventhub==2.5.0
azure-mgmt-hanaonazure==0.1.1
azure-mgmt-iotcentral==0.1.0
azure-mgmt-iothub==0.5.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-logic==3.0.0
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-managementgroups==0.1.0
azure-mgmt-managementpartner==0.1.0
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==1.0.0
azure-mgmt-monitor==0.5.2
azure-mgmt-msi==0.2.0
azure-mgmt-network==2.6.0
azure-mgmt-notificationhubs==2.0.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.1.0
azure-mgmt-powerbiembedded==2.0.0
azure-mgmt-rdbms==1.8.0
azure-mgmt-recoveryservices==0.3.0
azure-mgmt-recoveryservicesbackup==0.3.0
azure-mgmt-redis==5.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.2.1
azure-mgmt-resource==2.1.0
azure-mgmt-scheduler==2.0.0
azure-mgmt-search==2.0.0
azure-mgmt-servicebus==0.5.3
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.1.1
azure-mgmt-sql==0.9.1
azure-mgmt-storage==2.0.0
azure-mgmt-subscription==0.2.0
azure-mgmt-trafficmanager==0.50.0
azure-mgmt-web==0.35.0
azure-nspkg==3.0.2
azure-servicebus==0.21.1
azure-servicefabric==6.3.0.0
azure-servicemanagement-legacy==0.20.6
azure-storage-blob==1.5.0
azure-storage-common==1.4.0
azure-storage-file==1.4.0
azure-storage-queue==1.4.0
Babel==2.6.0
backcall==0.1.0
bambou==3.0.1
bcrypt==3.1.6
bleach==3.1.0
boto==2.49.0
boto3==1.6.6
botocore==1.9.6
bottle==0.12.16
bottle-sqlite==0.1.3
bugzilla-docstrings==0.0.1
cached-property==1.5.1
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.2
cfme-testcases==0.12
chardet==3.0.4
Click==7.0
cliff==2.14.1
cmd2==0.9.11
colorama==0.4.1
cryptography==2.6.1
cycler==0.10.0
dateparser==0.7.1
debtcollector==1.21.0
decorator==4.4.0
dectate==0.13
deepdiff==3.3.0
defusedxml==0.5.0
Deprecated==1.2.5
diaper==1.3
dictdiffer==0.8.0
distro==1.4.0
docker-py==1.10.6
docker-pycreds==0.4.0
docutils==0.14
dogpile.cache==0.7.1
dump2polarion==0.39.1
entrypoints==0.3
fauxfactory==3.0.2
flake8==3.7.7
flufl.enum==4.1.1
funcsigs==1.0.2
future==0.17.1
gevent==1.4.0
gitdb2==2.0.5
GitPython==2.1.11
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-compute-engine==2.8.13
greenlet==0.4.15
httplib2==0.12.1
humanfriendly==4.18
idna==2.8
importscan==0.1
inflection==0.3.1
ipaddress==1.0.22
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
iso8601==0.1.12
isodate==0.6.0
jedi==0.13.3
Jinja2==2.10.1
jira==2.0.0
jmespath==0.9.4
jsmin==2.2.2
jsonpatch==1.23
jsonpickle==1.1
jsonpointer==2.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
keystoneauth1==3.13.1
kubernetes==3.0.0
kwargify==2.2.0
layered-yaml-attrdict-config==18.12.3
lxml==4.3.3
manageiq-client==0.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
miq-version==0.1.11
mistune==0.8.4
mock==2.0.0
msgpack==0.6.1
msrest==0.6.6
msrestazure==0.6.0
multi-key-dict==2.0.3
multimethods.py==0.5.3
munch==2.3.2
navmazing==1.1.6
nbconvert==5.4.1
nbformat==4.4.0
netaddr==0.7.19
netifaces==0.10.9
notebook==5.7.8
ntlm-auth==1.3.0
oauth2client==4.1.3
oauthlib==3.0.1
openshift==0.3.4
openstacksdk==0.27.0
os-service-types==1.6.0
osc-lib==1.12.1
oslo.config==6.8.1
oslo.i18n==3.23.1
oslo.serialization==2.28.2
oslo.utils==3.40.3
ovirt-engine-sdk-python==4.3.0
packaging==19.0
pandocfilters==1.4.2
paramiko==2.6
parsedatetime==2.4
parso==0.4.0
pbr==5.1.3
pdfminer.six==20181108
pexpect==4.7.0
pickleshare==0.7.5
pika==1.0.1
Pillow==6.0.0
pluggy==0.6.0
polarion-docstrings==0.16.1
pre-commit==1.17.0
prettytable==0.7.2
progress==1.5
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psycopg2==2.8.2
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pycparser==2.17
pycryptodome==3.8.1
pycurl==7.43.0.2
pyflakes==2.1.1
pygal==2.4.0
PyGithub==1.43.7
Pygments==2.3.1
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyperclip==1.7.0
pytesseract==0.2.6
pytest==3.4.1
pytest-polarion-collect==0.15
python-box==3.2.4
python-bugzilla==2.2.0
python-cinderclient==4.1.0
python-dateutil==2.8.0
python-glanceclient==2.16.0
python-heatclient==1.17.0
python-ironicclient==2.7.0
python-jenkins==1.4.0
python-keystoneclient==3.19.0
python-novaclient==7.1.2
python-string-utils==0.6.0
python-swiftclient==3.7.0
pytz==2019.1
pyvcloud==19.1.2
pyvmomi==6.7.1.2018.12
py3winrm==0.0.1
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
redfish-client==0.1.0
reg==0.11
regex==2019.4.14
repoze.lru==0.7
requests==2.21.0
requests-ntlm==1.1.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
requestsexceptions==1.4.0
rfc3986==1.2.0
riggerlib==3.1.5
rsa==4.0
ruamel.yaml==0.15.92
s3transfer==0.1.13
scp==0.13.2
selenium==3.14.0
selenium-smart-locator==0.2.0
Send2Trash==1.5.0
sentaku==0.6.4
setuptools-scm==3.2.0
shyaml==0.6.1
simplejson==3.16.0
six==1.12.0
slumber==0.7.1
smmap2==2.0.5
snowballstemmer==1.2.1
sortedcontainers==2.1.0
Sphinx==1.3.5
sphinx-rtd-theme==0.4.3
SQLAlchemy==1.3.3
stevedore==1.30.1
suds-jurko==0.6
tabulate==0.7.7
taretto==0.5.5
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.24.2
virtualenv==16.4.3
vspk==5.3.2
wait-for==1.1.0
warlock==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.40.0
Werkzeug==0.15.2
widgetastic.core==0.39
widgetastic.patternfly==1.0.0
widgetsnbextension==3.4.2
wrapanapi==3.3.1
wrapt==1.11.1
xmltodict==0.12.0
yaycl==0.3.0
yaycl-crypt==0.4.0
